+++ README.md
@@ -4,65 +4,93 @@ This program is intended to export various AWS statistics as prometheus
 metrics. It is primarily intended to assist with billing. Currently the
 following metrics are exported:
 
-# EC2 Instance Counts
+# EC2 Instances
 
  - *aws_ec2_instances_count*: Count of istances
+ - *instancesNormalizationUnits*: Normalization units of istances
 
 The following labels are exposed:
 
- - *az*: availability zone
- - *instance_type*: type of instance
- - *lifecycle*: spot or scheduled instance
- - *groups*: sorted comma separated list of groups.
+ - *az*: Availability zone
+ - *family*: Instance family
+ - *groups*: [EC2-Classic only] sorted comma separated list of security groups
+ - *instance_type*: Type of instance
+ - *lifecycle*: spot, scheduled or normal instance
  - *owner_id*: The owner id
  - *requester_id*: The requester id (default to owner id if none is present)
+ - *units*: The normalization units of the instance
  - *aws_tag_*: Any tags passed in with the -instance-tags flag are added as labels
 
 # EC2 Reserved Instances
 Every set of instance reservations gets its own time series, this is intended to allow
 the end time of reserved intances to be tracked and potentially alerted upon.
 
- - *aws_ec2_reserved_instances_usage_price_dollars*: cost of reserved instance usage in dollars
- - *aws_ec2_reserved_instances_fixed_price_dollars*: fixed cost of reserved instance in dollars
- - *aws_ec2_reserved_instances_price_per_hour_dollars*: hourly cost of reserved instance in dollars
- - *aws_ec2_reserved_instances_count*: Number of reserved instances in this reservation
- - *aws_ec2_reserved_instances_start_time*: Start time of this reservation
- - *aws_ec2_reserved_instances_end_time*: End time of this reservation
+ - *aws_ec2_reserved_instances_fixed_unit_price*: Purchase price of the reservation per normalization unit in dollars
+ - *aws_ec2_reserved_instances_hourly_unit_price*: Hourly reservation reccuring charges per normalization unit in dollars
+ - *aws_ec2_reserved_instances_count*: Number of Reserved instances in this reservation
+ - *aws_ec2_reserved_instances_normalization_units_total*: Number of total normalization units in this reservation
 
 The following labels are exposed:
 
- - *id*: the reservation id
- - *az*: availability zone
- - *instance_type*: type of instance
- - *product*: The product description
- - *tenancy*:
+ - *az*: Availability zone
+ - *duration*: Duration of the reservation in seconds
+ - *end_date*: Date on which the reservation is retired
+ - *family*: Instance family
+ - *id*: The reservation id
+ - *instance_type*: Type of instance
  - *offer_type*:
+ - *product*: The product description
+ - *scope*: Region or Availability Zone
+ - *state*: The state of the Reserved Instance (payment-pending | active | payment-failed)
+ - *tenancy*: The tenancy of the instance (default | dedicated)
+ - *units*: The normalization units of the instance
+
+# EC2 Reserved Instances listed on the marketplace
+
+ - *aws_ec2_reserved_instances_listing_count*: Reserved instances listed on the market for a reservation
+
+The following labels are exposed:
+
+ - *az*: Availability zone
+ - *family*: Instance family
+ - *instance_type*: Type of instance
+ - *product*: The product description
+ - *scope*: Region or Availability Zone
+ - *state*: The state of the listed Reserved Instances
+ - *units*: The normalization units of the instance
 
 # EC2 Spot Instance Request
 
 Only fullfilled active spot instances requests are currently tracke
 
- - *aws\_ec2\_spot\_request\_count*: How active spot instances of a given type you have running
- - *aws\_ec2\_spot\_request\_bid\_price\_hourly\_dollars*: Your maximum bid price
- - *aws\_ec2\_spot\_request\_actual\_block\_price\_hourly\_dollars*: The price paid for limited duration spot instances
+ - *aws_ec2_spot_request_count*: How active spot instances of a given type you have running
+ - *aws_ec2_spot_request_bid_price_hourly_dollars*: Your maximum bid price
+ - *aws_ec2_spot_request_actual_block_price_hourly_dollars*: The price paid for limited duration spot instances
 
 The following labels are exposed:
 
  - *az*: availability zone
+ - *family*: Instance family
+ - *instance_profile*: The IAM instance profile
  - *instance_type*: type of instance
  - *product*: The product description
+ - *launch_group*: The Spot Instance launch group
+ - *persistence*: The type of Spot Instance request (one-time | persistent)
+ - *units*: The normalization units of the instance
 
 # EC2 Spot Instance Pricing
 
 Only prices for products that have been seen in spot instance requests are tracked.
 
- - *aws\_ec2\_spot\_price\_per\_hour\_dollars*: The current market price for a spot instance
+ - *aws_ec2_spot_price_per_hour_dollars*: The current market price for a spot instance
 
 The following labels are exposed:
 
  - *az*: availability zone
+ - *family*: Instance family
  - *instance_type*: type of instance
  - *product*: The product description
+ - *units*: The normalization units of the instance
 
 # Usage
 
@@ -76,10 +104,4 @@ The following labels are exposed:
   -instance-tags string
         comma seperated list of tag keys to use as metric labels
   -region string
-        the region to query (default "eu-west-1")
-
-# TODO
-
- - Add optional Push gateway support
- - Make tracking full Reserved instance tracking optional and pre-aggregate
-
+        the region to query (default "us-east-1")
